// biking/src/interfaces/BikeKeyFeatures.ts
interface IBikeKeyFeatures {
  id: number | undefined;
  name: string | undefined;
  instrumentConsole: string | undefined;
  gps: string | undefined;
  speedometer: string | undefined;
  odometer: string | undefined;
  fuelGauge: string | undefined;
  gearIndicator: string | undefined;
}

export default IBikeKeyFeatures;